+++ locals.tf
@@ -66,6 +66,7 @@ locals {
     dms_proxy_bucket                         = var.dms_proxy_bucket
     influx_db_hostname                       = var.is_influx_db_in_secondary_region ? var.influx_db_hostname : aws_timestreaminfluxdb_db_instance.influx_db[0].endpoint
     influx_db_api_token_arn                  = aws_secretsmanager_secret.influx_db_api_token.arn
+    iceberg_system_tables_db_name            = aws_glue_catalog_database.iceberg_system_tables_db.name
   }
 }